Move all the stuff to ahs_translation.py
