+++ src/braket/pennylane_plugin/translation.py
@@ -575,129 +575,3 @@ def translate_result(
         )
     else:
         return braket_result.get_value_by_result_type(translated)
-
-
-
-
-
-
-from numpy.typing import ArrayLike
-from typing import Callable
-from pennylane.pulse.hardware_hamiltonian import HardwarePulse
-from braket.timings.time_series import TimeSeries
-from braket.ahs.driving_field import DrivingField
-from braket.ahs.atom_arrangement import AtomArrangement
-from pennylane.pulse import ParametrizedEvolution
-
-from functools import partial
-
-def _convert_to_time_series(
-    pulse_parameter: Union[float, Callable],
-    time_points: ArrayLike,
-    scaling_factor: float = 1,
-):
-    """Converts pulse information into a TimeSeries
-
-    Args:
-        pulse_parameter(Union[float, Callable]): a physical parameter (pulse, amplitude
-            or frequency detuning) of the pulse. If this is a callalbe, it has already been
-            partially evaluated, such that it is only a function of time.
-        time_points(array): the times where parameters will be set in the TimeSeries, specified
-            in seconds
-        scaling_factor(float): A multiplication factor for the pulse_parameter
-            where relevant to convert between units. Defaults to 1.
-
-    Returns:
-        TimeSeries: a description of setpoints and corresponding times
-    """
-
-    ts = TimeSeries()
-
-    if callable(pulse_parameter):
-        # convert time to microseconds to evaluate (expected unit for the PL functions)
-        vals = [float(pulse_parameter(t * 1e6)) * scaling_factor for t in time_points]
-    else:
-        vals = [pulse_parameter * scaling_factor for t in time_points]
-
-    for t, v in zip(time_points, vals):
-        ts.put(t, v)
-
-    return ts
-
-def translate_pulse_to_driving_field(pulse: HardwarePulse, time_points: ArrayLike):
-    """Converts a ``HardwarePulse`` from PennyLane describing a global drive to a
-    ``DrivingField`` from Braket AHS
-
-    Args:
-        pulse[HardwarePulse]: a dataclass object containing amplitude, phase and frequency
-            detuning information
-        time_interval(array[float, float]): The start and end time for the applied pulse
-
-    Returns:
-        drive(DrivingField): the object representing the global drive for the
-            AnalogueHamiltonianSimulation object
-    """
-
-    # scaling factor for amp and frequency detuning converts from Mrad/s to rad/s
-    amplitude = _convert_to_time_series(pulse.amplitude, time_points, scaling_factor=1e6)
-    detuning = _convert_to_time_series(pulse.frequency, time_points, scaling_factor=1e6)
-    phase = _convert_to_time_series(pulse.phase, time_points)
-
-    drive = DrivingField(amplitude=amplitude, detuning=detuning, phase=phase)
-
-    return drive
-
-
-def _evaluate_pulses(ev_op: ParametrizedEvolution):
-    """Feeds in the parameters in order to partially evaluate the callables (amplitude,
-    phase and/or frequency detuning) describing the pulses, so they are only a function of time.
-    Saves the pulses on the device as `dev.pulses`.
-
-    Args:
-        ev_op(ParametrizedEvolution): the operator containing the pulses to be evaluated
-    """
-
-    params = ev_op.parameters
-    pulses = ev_op.H.pulses
-
-    evaluated_pulses = []
-    idx = 0
-
-    for pulse in pulses:
-        amplitude = pulse.amplitude
-        if callable(pulse.amplitude):
-            amplitude = partial(pulse.amplitude, params[idx])
-            idx += 1
-
-        phase = pulse.phase
-        if callable(pulse.phase):
-            phase = partial(pulse.phase, params[idx])
-            idx += 1
-
-        detuning = pulse.frequency
-        if callable(pulse.frequency):
-            detuning = partial(pulse.frequency, params[idx])
-            idx += 1
-
-        evaluated_pulses.append(
-            HardwarePulse(
-                amplitude=amplitude, phase=phase, frequency=detuning, wires=pulse.wires
-            )
-        )
-
-    return evaluated_pulses
-
-def _create_register(coordinates: List):
-    """Create an AtomArrangement to describe the atom layout from the coordinates in the
-    ParametrizedEvolution, and saves it as self._register
-
-    Args:
-        coordinates(List): a list of pairs [x, y] of coordinates denoting atom locations, in um
-    """
-
-    register = AtomArrangement()
-    for [x, y] in coordinates:
-        # PL asks users to specify in um, Braket expects SI units
-        register.add([x * 1e-6, y * 1e-6])
-
-    return register
